requirements.txt
django>4,<5.0
ipython
setuptools==70.2.0
django-markdownx==4.0.7
markdown-link-attr-modifier==0.2.1

# django-magicauth fork: Multiple Auth backend support
django-magicauth @ git+https://github.com/glibersat/django-magicauth@55eab0b91c95012944933b342748540b2c1b31bf

# Notifications Fork: Sites framework support
django-notifications-hq @ git+https://github.com/glibersat/django-notifications.git@feature/sites-framework

# Actstream fork: sites framework support
django-activity-stream @ git+https://github.com/glibersat/django-activity-stream.git@django4

django-multisite @ git+https://github.com/protoroto/django-multisite.git@feature/add-django4-support

django-dbtemplates @ git+https://github.com/jazzband/django-dbtemplates.git@master

django-cookie-consent @ git+https://github.com/jazzband/django-cookie-consent@master

# Django 4 patches
django-tagging@git+https://github.com/nextories/django-tagging.git@bug/usage_for_model_with_filters_none

django-ordered-model==3.7.4
django_otp==1.3.0

django-taggit==4.0.0 # move to 5.0
django-gravatar2
psycopg[binary]>3
django-leaflet
django-admin-csvexport==2.2
djangorestframework==3.15.2
djangorestframework-xml
djangorestframework-simplejwt
rest-framework-generic-relations
drf-spectacular[sidecar]
django-filter
dynamic-django-forms
django-crispy-forms==2.0
django-hijack==3.4.5
django-sesame[ua]==3.2.2
django-recaptcha==3.0.0
django-allauth==0.51.0
sentry-sdk
sib-api-v3-sdk
# STATIC FILES
django-sass-processor==1.4
libsass
django-compressor==4.4
multimethod==1.7
django-vite==3.0.3
django-phonenumber-field[phonenumbers]==6.3.0
django-watson==1.6.3
django-guardian==2.4.0
django-clone==5.3.3

django-reversion==5.0.2
django-reversion-compare==0.17.0

wagtail==6.1.3

celery==5.4.0
django-celery-results==2.5.1
redis==5.0.4

django-autoslug==1.9.9
metabase-api==3.4

# Django webhook fork
django-webhook @ git+https://github.com/etchegom/django-webhook.git@recoco

django-json-widget==2.0.1
